improve(melhorias): mudanca de algumas rotas, mudanca na forma de validacao de login para o access token e refresh token, inclusao de metodos para os jogos, melhorias de algumas rotas e tabelas, acerto das migrations.

+++ src/application/use-cases/articles/delete-article.use-case.ts
@@ -1,25 +1,49 @@
 import { HttpStatus, Injectable } from "@nestjs/common";
+import { ImageService } from "src/application/services/image.service";
 import { ApiResponseInterface } from "src/domain/interfaces/APIResponse.interface";
 import { ArticlesRepository } from "src/infrastructure/repositories/articles.repository";
 
 @Injectable()
 export class DeleteArticleUseCase {
 
-    constructor(private readonly articlesRepository: ArticlesRepository){}
+    constructor(
+        private readonly articlesRepository: ArticlesRepository,
+        private readonly imageService: ImageService
+    ){}
 
     async deleteArticle(id: number): Promise<ApiResponseInterface<number>> {
-        const deleteArticle = await this.articlesRepository.DeleteArticle(id);
-
-        if(deleteArticle !== 1){
-            return {
-                status: HttpStatus.NOT_FOUND,
-                message: "Ouve um erro ao remover o ApiUriTooLongResponse."
+        try{
+            const articleExists = await this.articlesRepository.findArticleById(id);
+            
+            if(!articleExists){
+                return {
+                    status: HttpStatus.NOT_FOUND,
+                    message: "Artigo não encontrado."
+                }
             }
-        }
 
-        return {
-            status: HttpStatus.OK,
-            message: "Artigo excluido com sucesso."
+            const deleteArticle = await this.articlesRepository.DeleteArticle(id);
+
+            if(deleteArticle !== 1){
+                return {
+                    status: HttpStatus.NOT_FOUND,
+                    message: "Ouve um erro ao remover o ApiUriTooLongResponse."
+                }
+            }
+
+            if(articleExists && articleExists.image){
+                await this.imageService.deleteImage(articleExists.image);
+            }
+
+            return {
+                status: HttpStatus.OK,
+                message: "Artigo excluido com sucesso."
+            }
+        }catch(error){
+            return {
+                status: HttpStatus.INTERNAL_SERVER_ERROR,
+                message: "Erro interno do servidor ao deletar o artigo."
+            }
         }
     }
 }
